# perf(utils/url): shorten mergePath and optimize for normal use cases.