@yusukebe 2025-02-11
Hi @usualoma 

It's very your-styled code!

> * The minified code should be small
> * It should be particularly fast when there are two arguments

I totally agree. I was also thinking #3909 is a rare case. I left one comment, but I think it's not a problem, so we can merge.